+++ libs/common/src/lib/interfaces/portfolio-public-details.interface.ts
@@ -1,10 +1,16 @@
 import { PortfolioPosition } from '@ghostfolio/common/interfaces';
 
 export interface PortfolioPublicDetails {
+  hasDetails: boolean;
   holdings: {
     [symbol: string]: Pick<
       PortfolioPosition,
-      'allocationCurrent' | 'countries' | 'name' | 'sectors' | 'value'
+      | 'allocationCurrent'
+      | 'countries'
+      | 'currency'
+      | 'name'
+      | 'sectors'
+      | 'value'
     >;
   };
 }